requirements.txt
streamlit
python-dotenv
langchain
langchain-google-generative-ai
pinecone
huggingface-hub
transformers
sentence-transformers
langchain-pinecone
langchain-community
pypdf